+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [v4.0.0] - 2020-04-08
+### Added
+`.removeEdges(key)` to remove all conncted edges in graph (directions in directed graph) from a vertex.
+
+#### Changed
+- `.removeEdge` & `.removeVertex` now return a boolean to indicated if something was removed.
+- `.addVertex` now returns the addded vertex. 
+
+### Removed
+- vertex none standard function `.serialize`.
+
+### Fixed
+- README
+- jsdoc
+
+## [v3.0.2] - 2020-03-06
+### Fixed
+- typos in readme
+
 ## [v3.0.1] - 2020-01-02
 ### Fixed
 - typos in readme